+++ example/models/d.go
@@ -1,3 +1,5 @@
+// Package models defines model types for example testing
+// +genx:doc
 package models
 
 import "github.com/xoctopus/sqlx/pkg/builder"