+++ pkg/builder/def_table_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/xoctopus/sqlx/pkg/builder"
 	"github.com/xoctopus/sqlx/pkg/frag"
 	"github.com/xoctopus/sqlx/pkg/frag/testutil"
+	"github.com/xoctopus/sqlx/testdata"
 )
 
 func TestTable(t *testing.T) {
@@ -137,7 +138,6 @@ func (WithAttrs) ColumnRel() map[string][]string {
 	}
 }
 
-/*
 func TestCatalog(t *testing.T) {
 	catalog := builder.CatalogFrom(
 		&testdata.User{},
@@ -171,4 +171,3 @@ func TestCatalog(t *testing.T) {
 		Expect(t, d.Desc, Equal(m.ColumnDesc()["ID"]))
 	})
 }
-*/